downloads-sorter/requirements.txt
anthropic
pdfplumber
python-docx
python-pptx
openpyxl
customtkinter
